palaeoverse: Prepare and Explore Data for Palaeobiological Analyses

Provides functionality to support data preparation and exploration for palaeobiological analyses, improving code reproducibility and accessibility. The wider aim of 'palaeoverse' is to bring the palaeobiological community together to establish agreed standards. The package currently includes functionality for data cleaning, binning (time and space), exploration, summarisation and visualisation. Reference datasets (i.e. Geological Time Scales https://stratigraphy.org/chart) and auxiliary functions are also provided. Details can be found in: Jones et al., (2023) doi: 10.1111/2041-210X.14099 .

See also

Author

Maintainer: Lewis A. Jones LewisA.Jones@outlook.com (ORCID)

Authors:

William Gearty willgearty@gmail.com (ORCID)

Bethany J. Allen Bethany.Allen@bsse.ethz.ch (ORCID)

Kilian Eichenseer kilian.eichenseer@gmail.com (ORCID)

Christopher D. Dean christopherdaviddean@gmail.com (ORCID)

Joseph T. Flannery-Sutherland jf15558@bristol.ac.uk (ORCID)

Other contributors:

Sofia Galvan sofia.galvan@uvigo.es (ORCID) [contributor]

Miranta Kouvari m.kouvari@ucl.ac.uk (ORCID) [contributor]

Pedro L. Godoy pedrolorenagodoy@gmail.com (ORCID) [contributor]

Cecily Nicholl cecily.nicholl@ucl.ac.uk (ORCID) [contributor]

Lucas Buffan lucas.l.buffan@gmail.com (ORCID) [contributor]

Erin M. Dillon emdillon23@gmail.com (ORCID) [contributor]

A. Alessandro Chiarenza a.chiarenza15@gmail.com (ORCID) [contributor]