Example dataset: Phanerozoic reefs from the PaleoReefs Database

A dataset of Phanerozoic reef occurrences from the PaleoReefs Database (PARED). This example dataset includes a subset of the available data from PARED, but can be used to demonstrate how the functions in the palaeoverse package might be applied.

Usage

reefs

Format

A data frame with 4363 rows and 14 variables:

r_number

Reference number given to the particular fossil reef in PARED

name

Reference name given to the particular fossil reef in PARED

formation

The geological formation to which the fossil reef belongs

system

The stratigraphic system to which the fossil reef belongs

series

The stratigraphic series to which the fossil reef belongs

interval

The stratigraphic interval to which the fossil reef belongs

biota_main

The main biota present within the fossil reef

biota_sec

The secondary biota present within the fossil reef

lng

The modern-day longitude of the fossil reef

lat

The modern-day latitude of the fossil reef

country

The country or ocean the fossil reef is located in

authors

The authors of the publication documenting the fossil reef

title

The title of the publication documenting the fossil reef

year

The year of the publication documenting the fossil reef

Source

Compiled by Lewis A. Jones. Downloaded on the 25th July 2022. doi:10.5281/zenodo.6037852

References

Kiessling, W. & Krause, M. C. (2022). PaleoReefs Database (PARED) (1.0) Data set. doi:10.5281/zenodo.6037852