Expand taxa to interval-level pseudo-occurrences

A function to generate interval-level pseudo-occurrences for taxa based on temporal ranges (e.g., the output of tax_range_time). While the resulting pseudo-occurrences should not be treated as equivalent to actual occurrence data (e.g., like that from the Paleobiology Database), such pseudo-occurrences may be useful for performing statistical analyses where the row representing a taxon must be replicated for each interval through which the taxon persisted.

Usage

tax_time_expand(
  taxdf,
  max_ma = "max_ma",
  min_ma = "min_ma",
  scale = "GTS2020",
  rank = "stage",
  ext_orig = TRUE
)

Arguments

taxdf

dataframe. A data.frame of taxa (such as that produced by tax_range_time) with columns for the maximum and minimum ages (FADs and LADs). Each row should represent a unique taxon. Additional columns may be included (e.g., taxon names, additional taxonomy, etc) and will be included in the returned data.frame. If required, numeric ages can be generated from interval names via the look_up function.

max_ma

character. The name of the column you wish to be treated as the maximum limit (FADs) of the age range (e.g., "max_ma").

min_ma

character. The name of the column you wish to be treated as the minimum limit (LADs) of the age range (e.g., "min_ma").

scale

character. Specify the desired geological timescale to be used, either "GTS2020" or "GTS2012".

rank

character. Specify the desired stratigraphic rank. Choose from: "stage", "epoch", "period", "era", and "eon".

ext_orig

logical. Should two additional columns be added to identify the intervals in which taxa originated and went extinct?

Value

A dataframe where each row represents an interval during which a taxon in the original user-supplied data persisted. The columns are identical to those in the user-supplied data with additional columns included to identify the intervals. If ext_orig is TRUE, two additional columns are added to identify in which intervals taxa originated and went extinct.

Developer(s)

William Gearty

Reviewer(s)

Lewis A. Jones

Examples

taxdf <- data.frame(name = c("A", "B", "C"),
                    max_ma = c(150, 60, 30),
                    min_ma = c(110, 20, 0))
tax_time_expand(taxdf)
#>    name max_ma min_ma   ext  orig interval_number     interval_name  rank
#> 1     C     30      0 FALSE FALSE               1        Meghalayan stage
#> 2     C     30      0 FALSE FALSE               2     Northgrippian stage
#> 3     C     30      0 FALSE FALSE               3      Greenlandian stage
#> 4     C     30      0 FALSE FALSE               4 Upper Pleistocene stage
#> 5     C     30      0 FALSE FALSE               5         Chibanian stage
#> 6     C     30      0 FALSE FALSE               6         Calabrian stage
#> 7     C     30      0 FALSE FALSE               7          Gelasian stage
#> 8     C     30      0 FALSE FALSE               8        Piacenzian stage
#> 9     C     30      0 FALSE FALSE               9          Zanclean stage
#> 10    C     30      0 FALSE FALSE              10         Messinian stage
#> 11    C     30      0 FALSE FALSE              11         Tortonian stage
#> 12    C     30      0 FALSE FALSE              12      Serravallian stage
#> 13    C     30      0 FALSE FALSE              13          Langhian stage
#> 14    B     60     20  TRUE FALSE              14       Burdigalian stage
#> 15    C     30      0 FALSE FALSE              14       Burdigalian stage
#> 16    B     60     20 FALSE FALSE              15        Aquitanian stage
#> 17    C     30      0 FALSE FALSE              15        Aquitanian stage
#> 18    B     60     20 FALSE FALSE              16          Chattian stage
#> 19    C     30      0 FALSE FALSE              16          Chattian stage
#> 20    B     60     20 FALSE FALSE              17          Rupelian stage
#> 21    C     30      0 FALSE  TRUE              17          Rupelian stage
#> 22    B     60     20 FALSE FALSE              18        Priabonian stage
#> 23    B     60     20 FALSE FALSE              19         Bartonian stage
#> 24    B     60     20 FALSE FALSE              20          Lutetian stage
#> 25    B     60     20 FALSE FALSE              21          Ypresian stage
#> 26    B     60     20 FALSE FALSE              22         Thanetian stage
#> 27    B     60     20 FALSE  TRUE              23         Selandian stage
#> 28    A    150    110  TRUE FALSE              31            Albian stage
#> 29    A    150    110 FALSE FALSE              32            Aptian stage
#> 30    A    150    110 FALSE FALSE              33         Barremian stage
#> 31    A    150    110 FALSE FALSE              34       Hauterivian stage
#> 32    A    150    110 FALSE FALSE              35       Valanginian stage
#> 33    A    150    110 FALSE FALSE              36        Berriasian stage
#> 34    A    150    110 FALSE  TRUE              37         Tithonian stage
#>      max_ma    mid_ma   min_ma duration_myr  font  colour
#> 1    0.0042   0.00210   0.0000       0.0042 black #FDEDEC
#> 2    0.0082   0.00620   0.0042       0.0040 black #FDECE4
#> 3    0.0117   0.00995   0.0082       0.0035 black #FEECDB
#> 4    0.1290   0.07035   0.0117       0.1173 black #FFF2D3
#> 5    0.7740   0.45150   0.1290       0.6450 black #FFF2C7
#> 6    1.8000   1.28700   0.7740       1.0260 black #FFF2BA
#> 7    2.5800   2.19000   1.8000       0.7800 black #FFEDB3
#> 8    3.6000   3.09000   2.5800       1.0200 black #FFFFBF
#> 9    5.3330   4.46650   3.6000       1.7330 black #FFFFB3
#> 10   7.2460   6.28950   5.3330       1.9130 black #FFFF73
#> 11  11.6300   9.43800   7.2460       4.3840 black #FFFF66
#> 12  13.8200  12.72500  11.6300       2.1900 black #FFFF59
#> 13  15.9700  14.89500  13.8200       2.1500 black #FFFF4D
#> 14  20.4400  18.20500  15.9700       4.4700 black #FFFF41
#> 15  20.4400  18.20500  15.9700       4.4700 black #FFFF41
#> 16  23.0300  21.73500  20.4400       2.5900 black #FFFF33
#> 17  23.0300  21.73500  20.4400       2.5900 black #FFFF33
#> 18  27.8200  25.42500  23.0300       4.7900 black #FEE6AA
#> 19  27.8200  25.42500  23.0300       4.7900 black #FEE6AA
#> 20  33.9000  30.86000  27.8200       6.0800 black #FED99A
#> 21  33.9000  30.86000  27.8200       6.0800 black #FED99A
#> 22  37.7100  35.80500  33.9000       3.8100 black #FDCDA1
#> 23  41.2000  39.45500  37.7100       3.4900 black #FDC091
#> 24  47.8000  44.50000  41.2000       6.6000 black #FDB482
#> 25  56.0000  51.90000  47.8000       8.2000 black #FCA773
#> 26  59.2000  57.60000  56.0000       3.2000 black #FDBF6F
#> 27  61.6000  60.40000  59.2000       2.4000 black #FEBF65
#> 28 113.0000 106.75000 100.5000      12.5000 black #CCEA97
#> 29 125.0000 119.00000 113.0000      12.0000 black #BFE48A
#> 30 129.4000 127.20000 125.0000       4.4000 black #B3DF7F
#> 31 132.6000 131.00000 129.4000       3.2000 black #A6D975
#> 32 139.8000 136.20000 132.6000       7.2000 black #99D36A
#> 33 145.0000 142.40000 139.8000       5.2000 black #8CCD60
#> 34 152.1000 148.55000 145.0000       7.1000 black #D9F1F7